code/requirements.txt
pandas
python-dotenv
openai
scikit-learn
numpy
sentence-transformers
